chore: clean up distributed mocks and unused files and folders
