--- __mocks__/components/SimpleMap/index.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const SimpleMap = () => {
-  return <div>A pretty map</div>;
-};
