--- __mocks__/grafana/app/core/app_events.ts
@@ -1,5 +1,0 @@
-export const appEvents = {
-  emit: () => {},
-};
-// eslint-disable-next-line no-restricted-syntax
-export default appEvents;
