--- cspell.config.json
@@ -1,5 +1,0 @@
-{
-  "ignorePaths": ["node_modules"],
-  "ignoreRegExpList": [".*"],
-  "words": ["grafana", "datasource", "datasources", "gifs", ".*"]
-}
