--- dummy.scss
@@ -1,1 +1,0 @@
-// this file exists solely to trick the spell check. Delete this as soon as possible
