+++ package.json
@@ -16,7 +16,6 @@
     "release:test": "release-it --dry-run --ci --no-git.requireCleanWorkingDir",
     "server": "docker compose up --build",
     "sign": "npx --yes @grafana/sign-plugin@latest",
-    "spellcheck": "cspell -c cspell.config.json \"**/*.{ts,tsx,js,go,md,mdx,yml,yaml,json,scss,css}\"",
     "test": "jest --all",
     "test:changed": "jest --watch --onlyChanged",
     "typecheck": "tsc --noEmit",
