--- src/__mocks__/raw-loader.js
@@ -1,3 +1,0 @@
-module.exports = {
-  process: (content) => 'module.exports = ' + JSON.stringify(content),
-};
