+++ src/test/mocks/components/SimpleMap.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+jest.mock(`components/SimpleMap`, () => ({
+  SimpleMap: () => <div>A pretty map</div>,
+}));
