+++ src/test/mocks/grafana/app/core/app_events.js
@@ -1,0 +1,3 @@
+module.exports = {
+  emit: () => {},
+};
